refactor: rename core journey package
